requirements.txt
requests
mysql-connector-python
pandas
tabulate